+++ tests/test_app_var.cpp
@@ -5,7 +5,7 @@
 TEST_CASE( "UNIT TEST: tracking class {Tracking}", "[core]" )
 {
     print_unit_msg("tracking class {Tracking}");
-    const unsigned par_per_dim = 8;
+    const size_t par_per_dim = 8;
     Tracking track(2, par_per_dim);
     std::vector<Particle_x<FTYPE_t>> particles;
     particles.resize(par_per_dim*par_per_dim*par_per_dim);
